+++ spec/support/cookbooks/box_sync_app_test/metadata.rb
@@ -1,0 +1,16 @@
+# Encoding: UTF-8
+#
+# rubocop:disable SingleSpaceBeforeFirstArg
+name             'box_sync_app_test'
+maintainer       'Jonathan Hartman'
+maintainer_email 'j@p4nt5.com'
+license          'apache2'
+description      'box_sync_app resource test wrapper'
+long_description 'box_sync_app resource test wrapper'
+version          '0.0.1'
+
+depends          'box-sync'
+
+supports         'mac_os_x'
+supports         'windows'
+# rubocop:enable SingleSpaceBeforeFirstArg