+++ spec/support/cookbooks/box_sync_app_test/recipes/remove.rb
@@ -1,0 +1,5 @@
+# Encoding: UTF-8
+
+box_sync_app 'default' do
+  action :remove
+end