+++ metadata.rb
@@ -7,7 +7,7 @@ maintainer_email 'j@p4nt5.com'
 license          'apache2'
 description      'Installs the Box Sync app'
 long_description 'Installs the Box Sync app'
-version          '0.1.1'
+version          '1.0.0'
 
 supports         'mac_os_x'
 supports         'windows'