+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Box Sync Cookbook CHANGELOG
 ===========================
 
+v?.?.? (????-??-??)
+-------------------
+
 v0.1.0 (2015-06-07)
 -------------------
 - Initial release! Supports OS X and Windows!