+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Box Sync Cookbook CHANGELOG
 
 v?.?.? (????-??-??)
 -------------------
+- Update provider resolver to use new style
 
 v0.1.0 (2015-06-07)
 -------------------